small tweaks

+++ src/main/java/com/peeko32213/unusualfishmod/client/renderer/RipperRenderer.java
@@ -27,7 +27,7 @@ public class RipperRenderer extends MobRenderer<Ripper, RipperModel<Ripper>> {
 		float f = 4.3F * Mth.sin(0.6F * p_116228_);
 		p_116227_.mulPose(Vector3f.YP.rotationDegrees(f));
 		if (!p_116226_.isInWater()) {
-			p_116227_.translate((double) 0.2F, (double) 0.1F, 0.0D);
+			p_116227_.translate(0.2F, 0.1F, 0.0D);
 			p_116227_.mulPose(Vector3f.ZP.rotationDegrees(90.0F));
 		}
 	}
